split out app
