+++ server/package.json
@@ -15,6 +15,7 @@
   "devDependencies": {
     "@types/dotenv": "^8.2.0",
     "@types/express": "^4.17.17",
+    "@types/jest": "^29.5.1",
     "@types/node": "^20.2.3",
     "@types/validator": "^13.7.17",
     "@typescript-eslint/eslint-plugin": "^5.59.7",
@@ -22,6 +23,8 @@
     "eslint": "^8.41.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^4.2.1",
+    "jest": "^29.5.0",
+    "supertest": "^6.3.3",
     "ts-node": "^10.9.1",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.0.4"
